Allow to pass multiple list of column coordinates
In case if we pass multiple areas it's a very reasonable use case to be able to pass column separators for each area. Tabula-java supports it natively and it works with submitting repeated column lists as java-native options. Additionally, it matches the order with the order of the areas.
It would be nice to have this feature wrapped into tabula-py.